chore: update dependencies and add .mergify.yml for automated merging (#105)
* chore: update dependencies and add .mergify.yml for automated merging

* fix: refactor project configuration file
