+++ .gitattributes
@@ -11,6 +11,7 @@
 /.github/workflows/release.yml linguist-generated
 /.github/workflows/upgrade.yml linguist-generated
 /.gitignore linguist-generated
+/.mergify.yml linguist-generated
 /.npmignore linguist-generated
 /.prettierignore linguist-generated
 /.prettierrc.json linguist-generated
