+++ .gitignore
@@ -46,6 +46,7 @@ node_modules
 .env.*
 .venv
 !/.github/workflows/build.yml
+!/.mergify.yml
 !/.github/workflows/upgrade.yml
 !/.github/pull_request_template.md
 !/.prettierignore
