+++ .npmignore
@@ -1,6 +1,7 @@
 # ~~ Generated by projen. To modify, edit .projenrc.ts and run "npx projen".
 /.projen/
 permissions-backup.acl
+/.mergify.yml
 /.prettierignore
 /.prettierrc.json
 /test/
