+++ .projen/files.json
@@ -9,6 +9,7 @@
     ".github/workflows/release.yml",
     ".github/workflows/upgrade.yml",
     ".gitignore",
+    ".mergify.yml",
     ".npmignore",
     ".prettierignore",
     ".prettierrc.json",
