+++ OSS_License_Summary.csv
@@ -6,14 +6,15 @@
 "(MIT OR CC0-1.0)","5"
 "(MIT OR GPL-3.0-or-later)","1"
 "0BSD","2"
-"Apache-2.0","150"
+"Apache-2.0","149"
 "BSD*","1"
 "BSD-2-Clause","17"
 "BSD-3-Clause","23"
 "CC-BY-3.0","1"
 "CC-BY-4.0","1"
 "CC0-1.0","2"
-"ISC","77"
-"MIT","667"
+"ISC","78"
+"MIT","675"
+"MIT-0","1"
 "Python-2.0","1"
 "UNLICENSED","1"
