+++ package-verification.json
@@ -1,11 +1,11 @@
 {
   "license": {
-    "package.json": "d244ff4d184a0c10ee104034bb116d720f3da40a89bc1457e4d90b721705ed06",
+    "package.json": "5e3c0f4eeea5509db7c8d7de30f81a924fe1d32b401d25bbeffa2b3fc9eb6eb5",
     "samples/cdk-ts-example/package.json": "54e2441af495ddf02e1e232fdd1ed4edc697c31fe03671776290a254e52ba2a6",
-    "packages/@cdklabs/cdk-cicd-wrapper/package.json": "1499d9e4859a8213890987c1378710fdfdd17d9b6de0d2bee45c4f2167b4d40b",
-    "packages/@cdklabs/cdk-cicd-wrapper-cli/package.json": "4adbdd47a202e7c72adc0b5d04e872d9e53518a63050cf0431931f85dd625fe4",
-    "packages/@cdklabs/cdk-cicd-wrapper-projen/package.json": "a3f820f251b73a3b599b159dd2d97733c5e05252e00f7c5845896e0909530b93",
+    "packages/@cdklabs/cdk-cicd-wrapper/package.json": "ab15e9ec5ef283bf9880008280f6e7c7e66a65cfbb5c41ed80d4bdeb586c7f8f",
+    "packages/@cdklabs/cdk-cicd-wrapper-cli/package.json": "158267d6b7d795be1abfb57e82da80eff72c8e45ad2a274c7bf83e370d7c2bfa",
+    "packages/@cdklabs/cdk-cicd-wrapper-projen/package.json": "9f36def0c9f26e41b9f48058a2a25e2b4e1a39f9d17bdd53c40d56687948a494",
     "projectList": "5790138ec76401840b75a0409a66d8e7e10445628567d18dbe5b2feedd20f463"
   },
-  "npm-lock-file": "cd2f6db304677e9bec24ec8464722eec0ed96b3ee50626948354d3f7b332b54f"
+  "npm-lock-file": "f7dcd2696e7c329cb51116b08613cafb171dd81f50641533c421f353deda4e69"
 }
