+++ packages/@cdklabs/cdk-cicd-wrapper-cli/package.json
@@ -37,18 +37,18 @@
     "eslint-config-prettier": "^9.1.0",
     "eslint-import-resolver-typescript": "^3.7.0",
     "eslint-plugin-import": "^2.31.0",
-    "eslint-plugin-prettier": "^5.2.1",
+    "eslint-plugin-prettier": "^5.2.3",
     "prettier": "^3.4.2",
-    "projen": "^0.91.5",
-    "typescript": "^5.7.2"
+    "projen": "^0.91.6",
+    "typescript": "^5.7.3"
   },
   "dependencies": {
-    "@aws-sdk/client-s3": "^3.721.0",
-    "@aws-sdk/credential-providers": "^3.721.0",
+    "@aws-sdk/client-s3": "^3.731.1",
+    "@aws-sdk/credential-providers": "^3.731.1",
     "@types/fs-extra": "^11.0.4",
     "@types/yargs": "^17.0.33",
     "csv": "^6.3.11",
-    "fs-extra": "^11.2.0",
+    "fs-extra": "^11.3.0",
     "globby": "11.1.0",
     "tslog": "^4.9.3",
     "yargs": "^17.7.2"
