+++ packages/@cdklabs/cdk-cicd-wrapper/package.json
@@ -36,8 +36,8 @@
     "organization": false
   },
   "devDependencies": {
-    "@aws-cdk/integ-runner": "2.149.0-alpha.0",
-    "@aws-cdk/integ-tests-alpha": "2.149.0-alpha.0",
+    "@aws-cdk/integ-runner": "^2.149.0-alpha.0",
+    "@aws-cdk/integ-tests-alpha": "^2.149.0-alpha.0",
     "@types/jest": "^29.5.14",
     "@types/node": "^18",
     "@typescript-eslint/eslint-plugin": "^7",
@@ -46,28 +46,29 @@
     "aws-cdk-lib": "2.149.0",
     "cdk-nag": "2.28.0",
     "commit-and-tag-version": "^12",
-    "constructs": "^10.0.0",
+    "constructs": "10.3.0",
     "copyfiles": "^2.4.1",
     "eslint": "^8",
     "eslint-config-prettier": "^9.1.0",
     "eslint-import-resolver-typescript": "^3.7.0",
     "eslint-plugin-import": "^2.31.0",
-    "eslint-plugin-prettier": "^5.2.1",
+    "eslint-plugin-prettier": "^5.2.3",
     "jest": "^29.7.0",
-    "jest-junit": "^15",
+    "jest-junit": "^16",
     "jsii": "~5.4.0",
     "jsii-diff": "^1.106.0",
     "jsii-pacmak": "^1.106.0",
     "jsii-rosetta": "~5.4.0",
     "prettier": "^3.4.2",
-    "projen": "^0.91.5",
+    "projen": "^0.91.6",
     "ts-jest": "^29.2.5",
-    "typescript": "^5.7.2"
+    "typescript": "^5.7.3"
   },
   "peerDependencies": {
     "aws-cdk-lib": "^2.149.0",
     "cdk-nag": "^2.28.0",
-    "constructs": "^10.0.0"
+    "cdk-pipelines-github": "^0.4.128",
+    "constructs": "^10.3.0"
   },
   "dependencies": {
     "@cloudcomponents/cdk-pull-request-approval-rule": "^2.4.0",
